order.Id use int64

+++ api/order.go
@@ -13,7 +13,7 @@ package api
 type Order struct {
 
 	// unique order id
-	Id int32 `json:"id,omitempty"`
+	Id int64 `json:"id,omitempty"`
 
 	// 'sell' or 'buy'
 	Side string `json:"side,omitempty"`
